:recycle: (pc-11984) deprecate BicolorLocationBuilding icon
