+++ src/features/cheatcodes/pages/AppComponents/Icons.tsx
@@ -20,7 +20,7 @@ import { BicolorBuilding } from 'ui/svg/icons/BicolorBuilding'
 import { BicolorConfidentiality } from 'ui/svg/icons/BicolorConfidentiality'
 import { BicolorEverywhere } from 'ui/svg/icons/BicolorEverywhere'
 import { BicolorFavorite } from 'ui/svg/icons/BicolorFavorite'
-import { BicolorLocationBuilding } from 'ui/svg/icons/BicolorLocationBuilding'
+import { BicolorLocationBuildingDeprecated } from 'ui/svg/icons/BicolorLocationBuilding_deprecated'
 import { BicolorLocationPointer } from 'ui/svg/icons/BicolorLocationPointer'
 import { BicolorLocationPointerDeprecated } from 'ui/svg/icons/BicolorLocationPointer_deprecated'
 import { BicolorLock } from 'ui/svg/icons/BicolorLock'
@@ -169,7 +169,10 @@ export const Icons: FunctionComponent = () => {
       <Icon name="BicolorEverywhere" component={BicolorEverywhere} isNew />
       <Icon name="BicolorLocationPointer" component={BicolorLocationPointer} isNew />
       <Icon name="BicolorLocationPointerDeprecated" component={BicolorLocationPointerDeprecated} />
-      <Icon name="BicolorLocationBuilding" component={BicolorLocationBuilding} />
+      <Icon
+        name="BicolorLocationBuildingDeprecated"
+        component={BicolorLocationBuildingDeprecated}
+      />
       <Icon name="NoOffer" component={NoOffer} />
       <Icon name="OfferDigital" component={OfferDigital} />
       <Icon name="OfferPhysical" component={OfferPhysical} />
