+++ src/features/search/components/locationChoice.utils.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { LocationType } from 'features/search/enums'
 import { useStagedSearch } from 'features/search/pages/SearchWrapper'
 import { AroundMeDeprecated } from 'ui/svg/icons/AroundMe_deprecated'
-import { BicolorLocationBuilding as LocationBuilding } from 'ui/svg/icons/BicolorLocationBuilding'
+import { BicolorLocationBuildingDeprecated as LocationBuilding } from 'ui/svg/icons/BicolorLocationBuilding_deprecated'
 import { BicolorLocationPointerDeprecated as LocationPointer } from 'ui/svg/icons/BicolorLocationPointer_deprecated'
 import { EverywhereDeprecated } from 'ui/svg/icons/Everywhere_deprecated'
 import { BicolorIconInterface } from 'ui/svg/icons/types'
