+++ src/libs/geolocation/components/WhereSection.tsx
@@ -13,7 +13,7 @@ import useOpenItinerary from 'libs/itinerary/useOpenItinerary'
 import { QueryKeys } from 'libs/queryKeys'
 import { Spacer } from 'ui/components/spacer/Spacer'
 import { ArrowNext } from 'ui/svg/icons/ArrowNext'
-import { BicolorLocationBuilding as LocationBuilding } from 'ui/svg/icons/BicolorLocationBuilding'
+import { BicolorLocationBuildingDeprecated as LocationBuilding } from 'ui/svg/icons/BicolorLocationBuilding_deprecated'
 import { Typo, ColorsEnum, getSpacing } from 'ui/theme'
 import { ACTIVE_OPACITY } from 'ui/theme/colors'
